Adding the skeleton for the new OOP-style filters
We won't use this in the first submission of the project, it's mainly just for learning purposes but I hope to use it in the later submissions.

I've used a lot of concepts from what we took in class like abstraction, polymorphism, inheritance, encapsulation, smart pointers, etc...

It's still uncomplete and unusable but I will continue it later.