More cleanup + fixed some potential infinites
There are still some remaining potential infinite loops but I am honestly too lazy to fix them. The root cause is that when we're doing input error-handling the while loop infinites if the user inputs a different datatype than the expected. To handle this, we should use cin.clear() and cin.ignore() stuff, will fix them later, for now will continue cleaning up the rest of the stuff as I am running out of time.