Setup the foundation for the GUI
With this commit, we're ready to work on the GUI, we're going to be using ImGUI + OpenGL3 + SDL2 as ImGui is bloat-free and enough for our vision. I mostly refactored the code to follow the structure I have in mind. I also deleted everything related to the CLI to keep it clean and replaced it with a very minimal main.cpp GUI that we can build upon.